Incorporate introduction/03 changes
Incorporate introduction/03 changes from the book. (As noted in the text of the
book, the behavior of hiding the element for the day when the day is clicked
instead of showing that day and hiding the others is not the intended final
behavior.

Refactored out the DayToggle class into a GenericToggle class in the src
directory and used it for both the button hide/show behavior and the date click
behavior at the top of the schedules/show page.

Also modify the view to provide the needed selectors to make the JavaScript
work.