Implement seat status toggle
Complete the Interactivity, State, and Hooks section of the React chapter. Modify
the Seat component to use React.useState() to allow the state of status to be
toggled on click of one of the seats and change the background color from white
to green.

Also got rid of the yarn-error.log and modified the .gitignore to make sure it
doesn't come back.