Add rotating chevron to favorites show-hide
Add a chevron svg that takes the place of the Show and Hide text for the
favorites list on the main page.

Add the css_flip_controller which is similar enough to the
css_toggle_controller to refactor out a css_parent class that both
extend.

Add transformations and transitions to the application.css file.