Add search to to root page
Make the search bar at the top of the page return search results on
input.

* Add Concert.search class method and use that to assign the @concerts
collection in concerts#index.
* Modify the concerts#index view to display search results within a
  Turbo Frame.
* Refactor the search bar from the schedules#show view into its own
  partial and create a form that submits the search to concerts#index.
* Create a Stimulus search_controller (pinned
  form-request-submit-polyfill to the import map to make it work with
  Safari)

As of now, there is no way to dismiss the results without refreshing
the page and the search can create too many meaningless round trips to
the server if the user types too fast.