Clear form and add debounce
Clear the search form and hide the results if the user clicks outside of
the form or the search results.

Also add a debounce function that starts a timeout and clears the
timeout if new input is reached before the timeout is reached. If new
input is received the old timer gets cleared and the new function
replaces it with the timer reset.