Use Stimulus to fetch server JSON
Create a SoldOutConcertsController on the Rails server side that only
has a show action and renders JSON with an array of sold out concert
IDs. Create a Stimulus sold_out_data_controller that fetches that
endpoint every minute and updates HTML data attributes of targets based
on results. This is a bit contrived to demonstrate the ability to do
this and wouldn't necessarily be the ideal implementation.

A Stimulus concert_controller updates the display of the UI when a
concert sells out. (Note that the reaction to a change in tickets
available is not yet implemented so if you change the tickets sold
number you would need to refresh the page to get updated values.)

Also pull in the ticket_tasks rake task that I forgot to include
earlier.