Use ActionCable with Turbo for favorites
Use ActionCable with Turbo for favorites and modify controllers for
concerts, favorites, and schedule days to use partials with a user local
variable instead of a current_user context variable.

Move the favorites create and destroy turbo_stream views to partials so
that we can invoke them for ActionCable.

Completes Turbo Streams and ActionCable section of Modern Front-End
Development for Rails chapter 10 Immediate Communication with
ActionCable