Add ActionCable channel for schedules
Add ActionCable channel for sold out concerts that replaces the polling
activity in the sold_out_data_controller on the Stimulus side. For now
the schedule_channel JavaScript file is not being used. The interaction
occurs in the Stimulus controller.

The broadcasting is currently only happening if the Rails
SoldOutConcertsController index action is hit, which is kind of
artificial.

Completes Stimulus and ActionCable section of chapter 10