Use Stimulus to implement calendar filter logic
Add calendar_controller to make the filtering for the schedule days at
the top of the main page work.

Modified the schedules/show and _schedule_day views to add new Stimulus
controller attirbutes and needed attributes to use the datasets of DOM
elements for filtering logic.

Also flipped the truth value on the css_toggle_controller so that true
means the classes are present and false means they are not present.

Completes Section Using Data Values for Logic in Chapter 11 Managing
State in Stimulus Code